Handle spec files using %patchN macros with rpm 4.19.90+
This doesn't solve the underlying issue of rpm in the pipeline not
matching the target release. It only works around one particular issue
by redefining problematic macros.

Note that the specfile with the redefinition is useless for building, it
would not actually apply the patches. There is a better way to redefine
the macro that would keep working, but it doesn't work with the specfile
library. See the linked centpkg PR for details of the better approach.

Relates: https://github.com/konflux-ci/rpmbuild-pipeline/issues/30
Relates: https://git.centos.org/centos/centpkg/pull-request/114