Grant Mock to read all sources (including subdirs)
We also need to set default:group:mock:r-x on the source directory
before we rsync sources from the Pod to the VM, so that any subdirectory
created later is also readable by Mock (in container).

We don't need to adjust permissions for the `mockbuilder` user
(UID=1000), because it already belongs to the `mock` group.

The `mock` group only needs read access to the sources; write
permissions are required only for the result directory.

Relates: #31